7.x islandora 2046: Allow Islandora Batch to be extended for more actions than just pure ingests (#1)
* First steps on abstract the repository action

* Using a get Method for action description

Suggested verb in past tense

* Remove writing back ingested object

To match qadan:7.x-ISLANDORA-2055

* Change queue primary Key

This will allow to have same PID in multiple different sets, just
restricting the same PID to exist multiple times in a single set

* Check if Object exists before calling ingest

This should have been there forever. Check if object is in place before
trying to ingest

* Combined Primary Key

This allows for same PID to exist in different Batch Sets. Quite useful
for so many reasons

* DCS